Fix: Creates Empty Author Folders if the user has turned that on in the UI
Readarr was missing the CreateEmptyAuthorFolders function which is similar to CreateEmptySeriesFolders from Sonarr.
All these changes were in the DiskScanService.cs file alone.
Updated the "public void Scan()" function to make it match Sonarr's mechanism for creating empty folders for authors/series.
Brought the Sonarr code over and changed "Series" references to "Author", also added the SetPermissions function that was a part of Sonarr's implementation.

This is part 2 of a set of changes to improve scan function to scan author folders and create them if they are missing.